fix some issue after move to aws

+++ services/core-service/src/services/aws-connector/aws-connector.service.ts
@@ -117,15 +117,21 @@ export class AwsConnectorService {
     } as S3.Types.ListObjectsRequest;
 
     const objects = await this.s3.listObjects(params).promise();
+    console.log("deleteFolder objects ", objects);
 
-    await this.s3
-      .deleteObjects({
-        Bucket: this.vultrUploadBucket,
-        Delete: {
-          Objects: objects.Contents?.map(({ Key }) => ({ Key })) ?? [],
-        },
-      })
-      .promise();
+    // make sure objects.Contents have objects, if not the function will be failed
+    if (objects.Contents.length > 0) {
+      await this.s3
+        .deleteObjects({
+          Bucket: this.vultrUploadBucket,
+          Delete: {
+            Objects: objects.Contents?.map(({ Key }) => ({ Key })) ?? [],
+          },
+        })
+        .promise();
+    }
+
+    console.log("deleteFolder keyFolder ", keyFolder);
 
     return this.s3
       .deleteObject({
